# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: Academia El Soto API
  description: REST API for Academia El Soto
  contact:
    name: Grupo 14
  version: 1.0.0
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/comments/{id}:
    get:
      tags:
      - comment-rest-controller
      operationId: getCommentById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentResponseDto"
    put:
      tags:
      - comment-rest-controller
      operationId: updateComment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentRequestDto"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentResponseDto"
    delete:
      tags:
      - comment-rest-controller
      operationId: deleteComment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
...
  /api/v1/courses:
    get:
      tags:
      - course-rest-controller
      operationId: getCourses
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCourseResponseDto"
    post:
      tags:
      - course-rest-controller
      operationId: createCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseRequestDto"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourseResponseDto"
...
  /api/v1/comments:
    get:
      tags:
      - comment-rest-controller
      operationId: getComments
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCommentResponseDto"
    post:
      tags:
      - comment-rest-controller
      operationId: createComment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentRequestDto"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentResponseDto"
...
  /api/v1/carts:
    get:
      tags:
      - cart-rest-controller
      operationId: getCarts
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCartResponseDto"
...
components:
  schemas:
...
    CourseRequestDto:
      type: object
      properties:
        courseName:
          type: string
          maxLength: 100
          minLength: 0
        teacher:
          type: string
          maxLength: 100
          minLength: 0
        price:
          type: integer
          format: int32
          minimum: 0
        description:
          type: string
          maxLength: 2000
          minLength: 0
        students:
          type: integer
          format: int32
          minimum: 0
        imageIds:
          type: array
          items:
            type: integer
            format: int64
      required:
      - courseName
      - description
      - price
      - students
      - teacher
    CourseResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        courseName:
          type: string
        teacher:
          type: string
        price:
          type: integer
          format: int32
        description:
          type: string
        students:
          type: array
          items:
            $ref: "#/components/schemas/UserBasicDto"
        imageIds:
          type: array
          items:
            type: integer
            format: int64
    UserBasicDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userName:
          type: string
        email:
          type: string
    CommentRequestDto:
      type: object
      properties:
        description:
          type: string
          maxLength: 500
          minLength: 0
        user:
          type: string
        courseId:
          type: integer
          format: int64
      required:
      - courseId
      - description
    CommentResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        user:
          type: string
        publicationDate:
          type: string
        courseId:
          type: integer
          format: int64
        courseName:
          type: string
...
    CartResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product:
          type: string
        price:
          type: integer
          format: int32
        userName:
          type: string
        courses:
          type: array
          items:
            $ref: "#/components/schemas/CourseResponseDto"
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageCourseResponseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CourseResponseDto"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageCommentResponseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CommentResponseDto"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageCartResponseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CartResponseDto"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT